Add Xilinx_32M register memory type
The register bank included implementations for memory primitives of
some older fpga architectures. RAM32M is present in Virtex-7 and on
paper, should be fairly good for our purposes. Sadly, it seems like
the synthesizer is quite good in this case and it is better to let it
infer the RAM. The synthesizer can create one less RAM primitive.